+++ crates/gdtoolkit-gdscript-lexer/src/token/common.rs
@@ -1,0 +1,4 @@
+/// Int type alias.
+pub type IntType = i64;
+/// Float type alias.
+pub type FloatType = f64;
