+++ crates/gdtoolkit-gdscript-lexer/src/token/keyword.rs
@@ -1,0 +1,49 @@
+use serde::{Deserialize, Serialize};
+
+/// Keyword.
+#[allow(missing_docs)]
+#[derive(Debug, PartialEq, Copy, Clone, Serialize, Deserialize)]
+pub enum Keyword {
+    And,
+    Assert,
+    Break,
+    Breakpoint,
+    Case,
+    Class,
+    ClassName,
+    Const,
+    Continue,
+    Elif,
+    Else,
+    Enum,
+    Export,
+    Extends,
+    For,
+    Func,
+    Get,
+    If,
+    In,
+    Is,
+    Master,
+    MasterSync,
+    Not,
+    Null,
+    OnReady,
+    Or,
+    Pass,
+    Puppet,
+    PuppetSync,
+    Remote,
+    RemoteSync,
+    Return,
+    SelfKw,
+    Set,
+    SetGet,
+    Signal,
+    Static,
+    Switch,
+    Tool,
+    Var,
+    While,
+    Yield,
+}
