+++ crates/gdtoolkit-gdscript-lexer/src/token/newline.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+/// Newline.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
+pub enum NewLine {
+    /// CRLF: "\r\n"
+    CrLf,
+    /// LF: "\n"
+    Lf,
+}
