+++ crates/gdtoolkit-gdscript-lexer/src/token/string.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+/// Quote mode.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
+pub enum QuoteMode {
+    /// Single quotes: '
+    Single,
+    /// Double quotes: "
+    Double,
+}
