+++ crates/gdtoolkit-gdscript-lexer/src/token/value.rs
@@ -1,0 +1,13 @@
+use serde::{Deserialize, Serialize};
+
+use super::{common::IntType, float::FloatRepr, string::QuoteMode};
+
+/// Value.
+#[allow(missing_docs)]
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub enum Value {
+    Boolean(bool),
+    String(String, QuoteMode),
+    Float(FloatRepr),
+    Int(IntType),
+}
